move temporary controllers to the root
